+++ Assets/scripts/OIPItemScript.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using UnityEngine.UI;
 
 public class OIPItemScript : MonoBehaviour {
-	public void updateContent(int value) {
+	public void updateContent(long value) {
 		if (this.GetComponentInChildren<Text> () != null)
 			this.GetComponentInChildren<Text> ().text = "" + value;
 	}
